pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-102"
version = "0.1.0"
description = ""
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
